PC-13 Added logger, adjusted file handling from request

+++ lesson_13_multithreading_and_multiprocessing/task_1_file_downloading/main.py
@@ -1,8 +1,14 @@
-import threading
+import logging
+import os
+from concurrent.futures import ThreadPoolExecutor
 from typing import List
+from urllib.error import HTTPError
+from urllib.parse import urlparse
 
 import requests
 
+
+logging.basicConfig(level=logging.DEBUG, format='%(asctime)s - %(levelname)s - %(message)s')
 
 def download_file(url: str) -> None:
     """
@@ -13,12 +19,30 @@ def download_file(url: str) -> None:
     try:
         response = requests.get(url)
         response.raise_for_status()
-        file_name = url.split('/')[-1]
+        if response.content is None:
+            logging.info(f'Content is empty.')
+            return
+        file_name = get_file_name_from_url(url)
+        if os.path.exists(file_name):
+            logging.info(f'File {file_name} already exists. Skip download.')
+            return
         with open(file_name, 'wb') as file:
             file.write(response.content)
-        print(f"File downloaded: {file_name}")
-    except Exception as e:
-        print("File downloading failed:", e)
+        logging.info(f'File {file_name} has been downloaded.')
+    except HTTPError as error:
+        logging.error("HTTP Error: {}".format(error))
+    except Exception as error:
+        logging.error(f"File downloading failed: {error}", )
+
+
+def get_file_name_from_url(url: str) -> str:
+    """
+    Get file name from url
+    :param url:
+    :return:
+    """
+    parsed_url = urlparse(url)
+    return os.path.basename(parsed_url.path)
 
 
 def downloading_files_by_threads(urls: List[str]) -> None:
@@ -27,14 +51,10 @@ def downloading_files_by_threads(urls: List[str]) -> None:
     :param urls:
     :return:
     """
-    threads = []
+    with ThreadPoolExecutor() as executor:
+        for url in urls:
+            executor.submit(download_file, url)
 
-    for url in urls:
-        thread = threading.Thread(target=download_file, args=(url,))
-        threads.append(thread)
-        thread.start()
-    for thread in threads:
-        thread.join()
 
 if __name__ == "__main__":
     urls = [
